Various small mods
commit a28dc5a34033212f9f17b32c7e1970f62dd4ac92 (TBAMax/master)
Author: Teemo Vaas <teemo@digiwood.ee>
Date:   Mon Mar 18 23:11:20 2024 +0200

Various small mods.

+++ PN5180.h
@@ -68,8 +68,8 @@ enum PN5180TransceiveStat {
 #define TX_RFOFF_IRQ_STAT   	(1<<8)  // RF Field OFF in PCD IRQ
 #define TX_RFON_IRQ_STAT    	(1<<9)  // RF Field ON in PCD IRQ
 #define RX_SOF_DET_IRQ_STAT 	(1<<14) // RF SOF Detection IRQ
-#define GENERAL_ERROR_IRQ_STAT 	(1<<17) // General error IRQ
-#define LPCD_IRQ_STAT 			(1<<19) // LPCD Detection IRQ
+#define GENERAL_ERROR_IRQ_STAT 	((uint32_t)1<<17) // General error IRQ
+#define LPCD_IRQ_STAT 		((uint32_t)1<<19) // LPCD Detection IRQ
 
 #define MIFARE_CLASSIC_KEYA 0x60  // Mifare Classic key A
 #define MIFARE_CLASSIC_KEYB 0x61  // Mifare Classic key B
